frontend/jupyter: move "Mode" (what is that?) to be part of the kernel menu, otherwise it wraps and is invisible + other cleanups
